Split models.py into models/ and package
