+++ app/models/__init__.py
@@ -1,0 +1,9 @@
+"""Import all models."""
+
+from .user import User
+from .about import About
+from .contact import Contact
+from .service import Service
+from .site import Site
+from .base_model import BaseModel
+from .service.environment_var import EnvironmentVar
